requirements-all.txt
bson==0.5.8
certifi==2020.4.5.1
chardet==3.0.4
dataclasses==0.6
defusedxml==0.6.0
Django==2.2.24
django-allauth==0.42.0
django-bulma-widget==1.2
django-crispy-forms==1.8.1
django-debug-panel==0.8.3
django-debug-toolbar==2.2
django-dotenv==1.4.2
django-grappelli==2.14.1
django-material==1.6.0
django-material-admin==1.6.42
django-pipeline==2.0.4
django-redis==4.11.0
djongo==1.3.0
idna==2.9
oauthlib==3.1.0
Pillow==8.2.0
psycopg2==2.9.1
pymongo==3.10.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
pywatchman==1.4.1
redis==3.4.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.2.4
urllib3==1.25.9
